moved filer files to new directory
